+++ R/util.R
@@ -1279,3 +1279,21 @@ summarise.metric_graph_data <- function(.data, ..., .include_graph_groups = FALS
 #' @export
 #' @usage lhs \%>\% rhs
 NULL
+
+
+#' @name summary.metric_graph
+#' @title Summary Method for \code{metric_graph} Objects
+#' @description Function providing a summary of several informations/characteristics of a metric graph object.
+#' @param object an object of class `metric_graph`.
+#' @param messages Should message explaining how to build the results be given for missing quantities?
+#' @param compute_characteristics Should the characteristics of the graph be computed?
+#' @param check_euclidean Check if the graph has Euclidean edges?
+#' @param check_distance_consistency Check the distance consistency assumption?#' 
+#' @param ... not used.
+#' @return An object of class \code{summary_graph_lme} containing information
+#' about a *metric_graph* object.
+#' @method summary metric_graph
+#' @export
+summary.metric_graph <- function(object, messages = FALSE, compute_characteristics = TRUE, check_euclidean = TRUE, check_distance_consistency = TRUE, ...){
+  object$summary(messages = messages, compute_characteristics = compute_characteristics, check_euclidean = check_euclidean, check_distance_consistency = check_distance_consistency)
+}